fix(declaration): rename categoryImportExport.ts to categoryFileHandler.ts
Avoids confusion with CategoryImportExport.tsx (React component) by
giving the utility module a more distinct name.
